[REFACTOR]: separate test file to based on context
repository, views, dan url
